requirements.txt
streamlit==1.26.0
streamlit-folium==0.13.0
Pillow==9.4.0
plotly==5.9.0
plotly.express==0.4.1
pandas==1.5.3
haversine==2.8.0
inflection==0.5.1
numpy==1.23.5
folium==0.14.0
